pyproject.toml
[tool.poetry]
name = "develop-cypers"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pytz = "^2019.3"
aiodns = "^2.0"
aiohttp = "^3.6"
bs4 = "^0.0.1"
cchardet = "^2.1"
discord = "^1.0"
pandas = "^1.0"
pycares = "^3.1"
request = "^2019.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
